ARROW-3701: [Gandiva] Workaround for jni JIT issue
- use the NativeLibrary.getInstance to force load the gandiva
  library with RT_GLOBAL.
