+++ java/gandiva/pom.xml
@@ -68,6 +68,11 @@
             <version>2.10</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>net.java.dev.jna</groupId>
+            <artifactId>jna</artifactId>
+            <version>4.5.0</version>
+        </dependency>
     </dependencies>
     <profiles>
         <profile>
